fix: four bugs in USB/buffer handling
Bug 1 (punker.h): unpack_u32 - add uint32_t cast before shifts to avoid
signed integer overflow UB (mirrors kernel fix c443be70).

Bug 2 (pcan_protocol.c): pcan_record_buffer_request - change 'next' from
uint8_t to uint16_t; uint8_t wraps at 256, corrupting the record buffer
at 28% capacity and resetting pos to 2.

Bug 3 (pcan_usb.c): pcan_usb_send_command_buffer / pcan_usb_send_data_buffer -
clear total_length on USBD_LL_Transmit failure; previously left set,
permanently blocking the endpoint until USB re-enumeration.

Bug 4 (punker.h): pack_u24 - fix operator precedence: (uint8_t*)(ptr)[N]
must be ((uint8_t*)(ptr))[N] to index after the cast, not before.

+++ Src/pcan_protocol.c
@@ -210,7 +210,7 @@ uint8_t *pcan_record_buffer_request( PCAN_RECORD_BUFFER_EX *prec, uint16_t size 
   if( unused < size )
   {
     /* align data to next slot bound */
-    uint8_t next = ( prec->pos + (BLOCK_SIZE-1) ) & (~((BLOCK_SIZE-1)));
+    uint16_t next = ( prec->pos + (BLOCK_SIZE-1) ) & (~((uint16_t)(BLOCK_SIZE-1)));
     if( (next+HEADER_SIZE) >= PCAN_MAX_RECORD_SIZE )
       return 0;
     memset( &prec->buffer[next], 0x00, BLOCK_SIZE );
